pyproject.toml
[tool.poetry]
name = "ocr"
version = "1.8.0"
description = ""
authors = ["lijun"]
readme = "README.md"
package-mode = false

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.bfsu.edu.cn/pypi/web/simple"
priority = "primary"

[tool.poetry.dependencies]
python = ">=3.10,<3.11.dev0"
absl-py = "0.11.0"
addict = "2.4.0"
anyconfig = "0.9.11"
cachetools = "4.2.4"
certifi = "2022.9.14"
chardet = "4.0.0"
click = "8.1.3"
cn2an = "0.5.17"
configobj = "5.0.6"
cycler = "0.11.0"
diskcache = "5.6.3"
easydict = "1.9"
Flask = "2.2.2"
func-timeout = "4.3.5"
future = "0.18.2"
gevent = "22.10.1"
gevent-websocket = "0.10.1"
google-auth = "1.24.0"
google-auth-oauthlib = "0.4.2"
greenlet = "1.1.3"
grpcio = "1.64.0"
idna = "2.10"
imageio = "2.9.0"
imgaug = "0.4.0"
itsdangerous = "2.0.0"
Jinja2 = "3.1.3"
kiwisolver = "1.4.4"
lmdb = "1.0.0"
lxml = "4.9.1"
Markdown = "3.3.3"
MarkupSafe = "2.1.5"
matplotlib = "3.8.4"
munch = "2.5.0"
natsort = "8.2.0"
networkx = "2.8.6"
numpy = "1.26.3"
oauthlib = "3.2.1"
opencv-python = "4.6.0.66"
packaging = "24.2"
pandas = "2.2.3"
pdfminer-six = "20231228"
pdfplumber = "0.11.4"
Pillow = "10.4.0"
proces = "0.1.2"
protobuf = "5.27.0"
pyasn1 = "0.4.8"
pyasn1-modules = "0.2.8"
pyclipper = "1.2.1"
pycryptodome = "3.15.0"
PyMuPDF = "1.20.2"
pyparsing = "3.0.9"
python-dateutil = "2.8.2"
python-docx = "1.1.0"
pytz = "2023.4"
PyWavelets = "1.4.1"
PyYAML = "6.0.2"
regex = "2022.9.13"
requests = "2.32.3"
requests-oauthlib = "1.3.1"
rsa = "4.6"
scikit-image = "0.19.3"
scipy = "1.13.1"
seaborn = "0.11.1"
setuptools = "63.4.1"
Shapely = "1.8.4"
six = "1.16.0"
sortedcontainers = "2.4.0"
thrift = "0.13.0"
tifffile = "2022.8.12"
tqdm = "4.67.1"
treelib = "1.6.1"
typing-extensions = "4.9.0"
urllib3 = "2.3.0"
Wand = "0.6.10"
Werkzeug = "2.3.8"
wheel = "0.37.1"
XlsxWriter = "1.3.7"
zhconv = "1.4.3"
megfile = "3.0.4"
loguru = "0.7.2"
Levenshtein = "0.25.1"
langchain = "0.0.347"
dynaconf = "3.2.5"
uvicorn = "0.29.0"
fastapi = "0.115.11"
redis = "5.0.4"
bs4 = "0.0.2"
python-multipart = "0.0.20"
pydantic = "2.9.2"
pydantic_core = "2.23.4"
pydantic-settings = "2.6.1"
tensorboard = "2.16.2"
transformers = "4.41.2"
ftfy = "6.3.1"
pypdfium2 = "4.30.0"
decorator = "5.1.1"
psutil = "6.1.0"
cloudpickle = "3.1.0"
tornado = "6.4.2"
optimum = "1.23.3"
cnstd = "1.2.4.2"
cnocr = "2.3.0.3"
pyspellchecker = "0.8.1"
pytest = "8.3.3"
ultralytics = "8.3.79"
doclayout_yolo = "0.0.3"
onnxruntime = "1.21.0"
rapid_table = "1.0.3"
gradio = "5.22.0"
gradio-pdf = "0.0.22"

onnxruntime-gpu = {version = "1.21.0", optional = true}
torch = {version = "2.1.0", optional = true}
torch-npu = {version = "2.1.0.post8", optional = true}
torchvision = {version = "0.16.0", optional = true}

[tool.poetry.extras]
gpu = ["onnxruntime-gpu"]
npu = ["torch", "torch-npu", "torchvision"]

[tool.poetry.scripts]
sdai-ocr = "SensedealImgAlg.WORKFLOW.OTHER.OCR.v0.OCRApiDeploy:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
